# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Actual Bridge API
  version: 1.0.0
  description: |
    Stable contract between the Android app and the Actual bridge.

    Conventions
    - All money values are integers in minor units (cents). Ledger amounts are signed:
      outflows negative, inflows positive. Budget amounts (`budgeted`, `totalBudgeted`)
      are the amount assigned, positive when money is assigned.
    - Months are `YYYY-MM`, dates are `YYYY-MM-DD`.
    - Within /v1 changes are additive only. Clients must ignore unknown fields and
      treat unknown enum values as `unknown`.
    - Non-idempotent writes require an `Idempotency-Key` header. Creates use a
      client-generated UUID `id` instead.
    - Errors are `application/problem+json` with a stable `code`.
servers:
  - url: https://budget-api.example.com
security:
  - bearer: []
...
paths:
...
  /v1/budgets/{budgetId}/transactions/batch:
    post:
      tags: [budgets]
      operationId: batchTransactions
      summary: Members. One change to many transactions (multi-select), or deleting them. Rows that can't take the change are skipped and listed.
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids: { type: array, minItems: 1, maxItems: 500, items: { type: string } }
                set:
                  type: object
                  description: Fields to set on every transaction. Send either set or delete.
                  properties:
                    categoryId: { type: [string, 'null'] }
                    accountId: { type: string }
                    cleared: { type: boolean }
                    notes: { type: [string, 'null'] }
                    date: { $ref: '#/components/schemas/Date' }
                    payeeId: { type: [string, 'null'] }
                delete: { type: boolean }
      responses:
        '200':
          description: What changed
          content:
            application/json:
              schema:
                type: object
                required: [updated, deleted, skipped]
                properties:
                  updated: { type: integer }
                  deleted: { type: integer }
                  skipped:
                    type: array
                    items:
                      type: object
                      required: [id, reason]
                      properties:
                        id: { type: string }
                        reason: { type: string, enum: [not_found, split, split_part, transfer] }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/review:
    get:
      tags: [budgets]
      operationId: getReviewInbox
      summary: "This person's review inbox: transactions added since they last caught up, plus anything uncategorized, minus what they've reviewed. Each person has their own; your own entries and edits count as reviewed."
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200':
          description: Newest first
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewInbox' }
    post:
      tags: [budgets]
      operationId: markReviewed
      summary: Mark transactions reviewed, or everything (`all`, which also moves the watermark to now).
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids: { type: array, minItems: 1, maxItems: 500, items: { type: string } }
                all: { type: boolean, const: true }
      responses:
        '200':
          description: How many are left
          content:
            application/json:
              schema: { type: object, required: [remaining], properties: { remaining: { type: integer } } }
  /v1/budgets/{budgetId}/transfers/matches:
    get:
      tags: [budgets]
      operationId: transferMatches
      summary: "Pairs that look like one payment seen from both accounts (a credit card payment leaving checking and arriving on the card): same amount, opposite signs, different accounts, within five days, neither already a transfer. Last 120 days, newest first. `confident` means neither side had another candidate."
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      responses:
        '200':
          description: Pairs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransferMatches' }
  /v1/budgets/{budgetId}/transfers/candidates:
    get:
      tags: [budgets]
      operationId: transferCandidates
      summary: "The other side of a transaction, to link by hand: transactions in other accounts that aren't transfers yet. The same amount going the other way first (exact), then the closest dates. Without q, two months either side; q searches payee, notes and amount over a year either side."
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - { name: transactionId, in: query, required: true, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 30 } }
      responses:
        '200':
          description: Candidates
          content:
            application/json:
              schema:
                type: object
                required: [transaction, items]
                properties:
                  transaction: { $ref: '#/components/schemas/Transaction' }
                  items:
                    type: array
                    items:
                      type: object
                      required: [transaction, exact, days]
                      properties:
                        transaction: { $ref: '#/components/schemas/Transaction' }
                        exact: { type: boolean, description: The same amount going the other way }
                        days: { type: integer }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/transfers/link:
    post:
      tags: [budgets]
      operationId: linkTransfer
      summary: Make the pair one transfer (what Actual's "Make transfer" does). Both stop counting as spending or income; on-budget pairs lose their categories.
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransferPairRequest' }
      responses:
        '200':
          description: Both sides, now linked
          content:
            application/json:
              schema:
                type: object
                required: [from, to]
                properties:
                  from: { $ref: '#/components/schemas/Transaction' }
                  to: { $ref: '#/components/schemas/Transaction' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/transfers/dismiss:
    post:
      tags: [budgets]
      operationId: dismissTransfer
      summary: Not the same payment; stop suggesting this pair.
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransferPairRequest' }
      responses:
        '204': { description: Dismissed }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/transfers/settings:
    put:
      tags: [budgets]
      operationId: setTransferSettings
      summary: With auto on, every bank sync links the confident pairs by itself.
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [auto], properties: { auto: { type: boolean } } }
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema: { type: object, required: [auto], properties: { auto: { type: boolean } } }
        '403': { $ref: '#/components/responses/Problem' }
...
  /v1/budgets/{budgetId}/transactions:
    get:
      tags: [budgets]
      operationId: listTransactions
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - { name: accountId, in: query, schema: { type: string } }
        - { name: categoryId, in: query, schema: { type: string } }
        - { name: payeeId, in: query, schema: { type: string } }
        - { name: groupId, in: query, description: Transactions in any category of this group, schema: { type: string } }
        - { name: since, in: query, schema: { type: string, format: date } }
        - { name: until, in: query, schema: { type: string, format: date } }
        - { name: q, in: query, description: Search payee, notes and category name, schema: { type: string, maxLength: 100 } }
        - { name: uncategorized, in: query, description: Only on-budget transactions that still need a category, schema: { type: boolean } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Newest first, splits grouped under their parent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionPage' }
        '423': { $ref: '#/components/responses/Problem' }
    post:
      tags: [budgets]
      operationId: createTransaction
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewTransaction' }
      responses:
        '201':
          description: Created. Re-sending the same `id` returns the existing transaction with 200.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '200':
          description: Already existed (idempotent replay)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/transactions/{transactionId}:
    parameters:
      - $ref: '#/components/parameters/BudgetId'
      - { name: transactionId, in: path, required: true, schema: { type: string } }
    get:
      tags: [budgets]
      operationId: getTransaction
      responses:
        '200':
          description: The transaction, with splits grouped
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [budgets]
      operationId: updateTransaction
      description: |
        Partial update. Changing the payee to a transfer payee turns it into a transfer.
        `subtransactions` replaces all splits (capability `transactions.splits`); send an
        existing split's `id` to keep it, omit ids for new splits, `[]` to unsplit.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionPatch' }
      responses:
        '200':
          description: The updated transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '501': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [budgets]
      operationId: deleteTransaction
      description: Also removes a transfer's other side and a split's children.
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    BudgetId:
      name: budgetId
      in: path
      required: true
      description: Actual sync id (the budget's `groupId`), stable across devices
      schema: { type: string }
...
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    TransferPairRequest:
      type: object
      required: [fromId, toId]
      properties:
        fromId: { type: string, description: The side money left }
        toId: { type: string, description: The side money arrived }
    TransferMatches:
      type: object
      required: [pairs, auto]
      properties:
        auto: { type: boolean }
        pairs:
          type: array
          items:
            type: object
            required: [from, to, days, confident]
            properties:
              from: { $ref: '#/components/schemas/Transaction' }
              to: { $ref: '#/components/schemas/Transaction' }
              days: { type: integer, description: Days between the two dates }
              confident: { type: boolean }
...
    Money:
      type: integer
      format: int64
      description: Minor units (cents)
...
    Date:
      type: string
      pattern: '^\d{4}-\d{2}-\d{2}$'
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: Stable machine-readable code, e.g. `unauthorized`, `forbidden`, `not_found`, `validation`, `feature_unavailable`, `actual_unavailable`, `budget_encrypted`, `bank_sync_failed`, `update_failed`, `conflict`, `rate_limited`
        detail: { type: string }
...
    Transaction:
      type: object
      required: [id, accountId, date, amount, cleared, reconciled, isParent, subtransactions]
      properties:
        id: { type: string }
        accountId: { type: string }
        date: { $ref: '#/components/schemas/Date' }
        amount: { $ref: '#/components/schemas/Money' }
        payeeId: { type: [string, 'null'] }
        payeeName: { type: [string, 'null'] }
        categoryId: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        cleared: { type: boolean }
        reconciled: { type: boolean }
        transferId: { type: [string, 'null'] }
        isParent: { type: boolean }
        parentId: { type: [string, 'null'] }
        subtransactions:
          type: array
          items: { $ref: '#/components/schemas/Transaction' }
    TransactionPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Transaction' }
        nextCursor: { type: [string, 'null'] }
        runningBalances:
          type: [array, 'null']
          items: { $ref: '#/components/schemas/Money' }
          description: The account's balance after each item, in order. Only when listing one account with no other filters.
...
    ReviewInbox:
      type: object
      required: [items, total, more, since]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Transaction' }
        total: { type: integer, description: How many need review (up to 500) }
        more: { type: boolean, description: More than 500 }
        since: { type: string, format: date-time, description: When this person last caught up }
    NewTransaction:
      type: object
      required: [id, accountId, date, amount]
      properties:
        id:
          type: string
          format: uuid
          description: Client-generated. Makes the create idempotent for outbox replays.
        accountId: { type: string }
        date: { $ref: '#/components/schemas/Date' }
        amount: { $ref: '#/components/schemas/Money' }
        payeeId: { type: string }
        payeeName: { type: string, description: Used when payeeId is absent; creates the payee if needed }
        categoryId: { type: string }
        notes: { type: string }
        cleared: { type: boolean }
        subtransactions:
          type: array
          items: { $ref: '#/components/schemas/SplitInput' }
    SplitInput:
      type: object
      required: [amount]
      properties:
        id: { type: string, description: An existing split's id (PATCH only) }
        amount: { $ref: '#/components/schemas/Money' }
        categoryId: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
    TransactionPatch:
      type: object
      minProperties: 1
      properties:
        accountId: { type: string }
        date: { $ref: '#/components/schemas/Date' }
        amount: { $ref: '#/components/schemas/Money' }
        payeeId: { type: [string, 'null'] }
        payeeName: { type: string, description: Used when payeeId is absent; creates the payee if needed }
        categoryId: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        cleared: { type: boolean }
        subtransactions:
          type: array
          items: { $ref: '#/components/schemas/SplitInput' }
...